backend/requirements.txt
asgiref==3.8.1
cfgv==3.4.0
coverage==7.7.0
distlib==0.3.9
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
filelock==3.18.0
identify==2.6.9
iniconfig==2.0.0
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.10.0
PyYAML==6.0.2
ruff==0.11.0
sqlparse==0.5.3
virtualenv==20.29.3
